requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
asarPy==1.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Cython==3.0.11
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.67.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5.dev0
Markdown==3.5.2
MarkupSafe==3.0.1
meson==1.5.2
multidict==6.1.0
numpy==2.1.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.27.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyGObject==3.50.0
PyJWT==2.9.0
pyparsing==3.2.0
pytesseract==0.3.13
pytest==8.3.3
requests==2.32.3
rsa==4.9
setuptools==75.2.0.post20241017
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.36
tqdm==4.66.5
twilio==9.3.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
yarl==1.16.0
